extracted terminal utils into seperate module
